攻击, 攻击封包 受伤 受伤封包ok

+++ DB/src/main/java/com/zhaoxiaodan/mirserver/db/entities/StdMonster.java
@@ -17,7 +17,7 @@ public class StdMonster {
 	public boolean undead;
 	public byte coolEye;
 	public int exp;
-	public int hp;
+	public short hp;
 	public int mp;
 	public int ac;
 	public int mac;
